Move aggregate-field-vm content to tree-field
