+++ src/ggrc-client/js/components/tree/templates/tree-field.stache
@@ -3,8 +3,18 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 }}
 
-<div rel="tooltip" data-original-title="{{resultStr}}">
+{{#if showTooltip}}
+  <div
+    rel="tooltip"
+    data-original-title="{{tooltipContent}}"
+  >
+    {{#items}}
+      <div class="tree-field__item">{{.}}</div>
+    {{/items}}
+  </div>
+{{else}}
   {{#items}}
     <div class="tree-field__item">{{.}}</div>
   {{/items}}
-</div>
+{{/if}}
+
